fix: move types to deferred executor
